feat(ipc): switch responses to NDJSON with stable error codes
Requests stay bare-verb (PING, STATUS, DUMP …); responses become
newline-delimited JSON objects carrying a v:1 protocol version and a
type discriminator (pong/status/sessions/ok/ok_path/error), per
docs/decisions/20260705-ndjson-ipc.md. This replaces the ad-hoc
key:value,key:value line tokens that broke on commas/colons inside app
names or messages.

IPCResponse.err(String) becomes .error(code:, message:) so consumers
(CommandRouter, DaemonSession) match on a stable code instead of
substring-matching the human message. DUMP SESSION ids move to UInt64
to match SessionRegistry's stable ids.

+++ Sources/ManbokCore/IPC/IPCCommand.swift
@@ -23,7 +23,7 @@ public enum IPCCommand: Equatable, Sendable {
     case status
     case sessions
     case stop
-    case dump(minutes: Int?, sessionId: Int?)
+    case dump(minutes: Int?, sessionId: UInt64?)
 
     /// Wire form without trailing newline (caller adds `\n` when sending).
     public var wireLine: String {
@@ -65,7 +65,7 @@ public enum IPCCommand: Equatable, Sendable {
         case "STOP":
             return parts.count == 1 ? .stop : nil
         case "DUMP":
-            if parts.count == 3, parts[1].uppercased() == "SESSION", let id = Int(parts[2]), id >= 1 {
+            if parts.count == 3, parts[1].uppercased() == "SESSION", let id = UInt64(parts[2]), id >= 1 {
                 return .dump(minutes: nil, sessionId: id)
             }
             switch parts.count {
